refactor: UI 컴포넌트 패키지 구조 정리 및 FAB 네이밍 변경
- `scaffold` 패키지 하위에 있던 `topbar`, `bottombar` 패키지를 `core.ui` 바로 아래로 이동하여 계층 단순화
- Floating Action Button 관련 컴포넌트들을 `core.ui.button.FAB` 패키지로 통합 및 이동
- `PenFAB`의 명칭을 `PenFloatingActionButton`으로 변경하여 명확성 및 통일성 강화
- 패키지 구조 및 클래스명 변경에 따른 프로젝트 내 임포트 경로 업데이트
- `HomeTabScreen`에서 불필요한 주석 제거 및 코드 정리
