+++ app/src/main/java/com/afternote/afternote_fe/navigation/AppNavigation.kt
@@ -12,7 +12,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.afternote.afternote_fe.screen.HomeTabScreen
 import com.afternote.core.ui.Route
-import com.afternote.core.ui.scaffold.bottombar.BottomBar
+import com.afternote.core.ui.bottombar.BottomBar
 import com.afternote.core.ui.theme.AfternoteDesign
 import com.afternote.feature.afternote.presentation.author.navigation.AfternoteNavGraphParams
 import com.afternote.feature.afternote.presentation.author.navigation.afternoteNavGraph
