+++ core/ui/src/main/kotlin/com/afternote/core/ui/topbar/DetailTopBar.kt
@@ -1,4 +1,4 @@
-package com.afternote.core.ui.scaffold.topbar
+package com.afternote.core.ui.topbar
 
 import android.content.res.Configuration
 import androidx.compose.foundation.layout.Arrangement
